requirements-dev.txt
# Lambda runtime dependencies
boto3>=1.34.0
PyJWT>=2.8.0

# Test dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto[dynamodb,ses,ssm]>=5.0.0
freezegun>=1.2.0

# Linting
ruff>=0.1.0
